Refactor search pages and link to use dynamic routing
Using next.js dynamic routing (See: https://nextjs.org/docs#dynamic-routing) We can create pages with dynamic routes e.g /search/all
