+++ app/pages/search/all.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../components/Layout';
+import Layout from '../../components/Layoutout';
 
 function SearchAll() {
   return <Layout title="Search">Search All</Layout>;
