+++ app/pages/search/content.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../components/Layout';
+import Layout from '../../components/Layoutout';
 
 function SearchContent() {
   return <Layout title="Search">Search Content</Layout>;
