+++ app/pages/search/groups.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../components/Layout';
+import Layout from '../../components/Layoutout';
 
 function SearchGroups() {
   return <Layout title="Search">Search Groups</Layout>;
