+++ app/pages/search/users.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import Layout from '../components/Layout';
+import Layout from '../../components/Layoutout';
 
 function SearchUsers() {
   return <Layout title="Search">Search Users</Layout>;
